+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Better handling of virtual files
 - Silence an error that is thrown when restarting dprint
 - Improve verbose logging in the console
+- Add a listener to detect config changes, note this only detects changes made inside IntelliJ
 
 ## [0.3.5]
 - Fix issue when performing code refactoring